Struct c_linked_list::CLinkedList

pub struct CLinkedList<T, P, F: Fn(&T) -> P> { /* fields omitted */ }

A linked list for wrapping a C linked list.

This struct is created by the from_const_ptr method and from_mut_ptr method from a C linked list that elements are linked by immutable pointers and mutable pointers, respectively. See their documentation for more.

Methods

impl<T, F> CLinkedList<T, *const T, F> where
F: Fn(&T) -> *const T,

pub fn from_const_ptr(head: *const T, next: F) -> Option<Self>

Creates a CLinkedList by wrapping a C linked list. head points to the head element of the list or is NULL for a list of length 0. next is a function that takes an element and returns an immutable raw pointer to the next element.

pub fn len(&self) -> usize

Returns the length of the CLinkedList.

pub fn contains(&self, x: &T) -> bool where
T: PartialEq<T>,

Returns true if the CLinkedList contains an element equal to the given value.

impl<T, F> CLinkedList<T, *mut T, F> where
F: Fn(&T) -> *mut T,

pub fn from_mut_ptr(head: *mut T, next: F) -> Option<Self>

Creates a CLinkedList by wrapping a C linked list. head points to the head element of the list or is NULL for a list of length 0. next is a function that takes an element and returns a mutable raw pointer to the next element.

pub fn iter_mut(&mut self) -> IterMut<T, *mut T, F>

Provides a forward iterator with mutable references.

pub fn len(&self) -> usize

Returns the length of the CLinkedList.

pub fn contains(&self, x: &T) -> bool where
T: PartialEq<T>,

Returns true if the CLinkedList contains an element equal to the given value.

pub fn front_mut(&mut self) -> Option<&mut T>

Provides a mutable reference to the front element, or None if the list is empty.

impl<T, P, F> CLinkedList<T, P, F> where
F: Fn(&T) -> P,

pub fn iter(&self) -> Iter<T, P, F>

Provides a forward iterator.

pub fn is_empty(&self) -> bool

Returns true if the CLinkedList is empty.

pub fn front(&self) -> Option<&T>

Provides a reference to the front element, or None if the list is empty.