Struct c_linked_list::CLinkedListConst

pub struct CLinkedListConst<T, N: Fn(&T) -> *const T> {
    // some fields omitted
}

Wraps a C linked list comprised of const pointers between nodes.

Methods

impl<'a, T: 'a, N: Fn(&T) -> *const T + 'a> CLinkedListConst<T, N>

unsafe fn from_ptr(head: *const T, next: N) -> CLinkedListConst<T, N>

Construct a CLinkedListConst by wrapping a C linked list. head points to the head element of the list or is NULL for a list of length 0. next is a function that takes a node and returns a pointer to the next element.

Example

To wrap this C type.

struct LinkedListNode {
    int value;
    const struct LinkedListNode *next;
};

Call this function as CLinkedListConst::from_ptr(ptr_to_head, |n| n.next).

Unsafety

This function is unsafe because it is up to the caller to ensure that head is valid.

fn iter(&'a self) -> CLinkedListConstIter<'a, T, N>

Iterate over the linked list, returning immutable references to the nodes of the list.

fn is_empty(&self) -> bool

Returns true if the list is empty.

fn len(&self) -> usize

Calculates the length of the list. This is an O(n) operation.

fn front(&self) -> Option<&T>

Provides a reference to the front element in the list, or None if the list is empty.