Struct c_linked_list::CLinkedListMut

pub struct CLinkedListMut<T, N: FnMut(&mut T) -> *mut T> {
    // some fields omitted
}

Wraps a C linked list comprised of mutable pointers between nodes.

Methods

impl<'a, T: 'a, N: FnMut(&mut T) -> *mut T + 'a> CLinkedListMut<T, N>

fn from_ptr(head: *mut T, next: N) -> CLinkedListMut<T, N>

Construct a CLinkedListMut by wrapping a C linked list. head points to the head element of the list and can be NULL for a list of length 0. next is a function that takes a node and returns a pointer to the next element.

Example

To wrap this C type.

struct LinkedListNode {
    int value;
    struct LinkedListNode *next;
};

Call this function as CLinkedListMut::from_ptr(ptr_to_head, |n| n.next).

fn iter(&'a mut self) -> CLinkedListMutIter<'a, T, N>

Iterate over the linked list, returning mutable references to the nodes of the list.